requirements.txt
pandas==2.2.0
python-dotenv==1.0.1
colorama==0.4.6
pytest==8.0.0
pytest-cov==4.1.0